Polished Weapon Mechanics Section
Increased size and window of time for collision detection. Removed one of the trails from sword. Disabled the low hammer attack because it needs to have its reach increased to be easier to hit an enemy with. Adjusted controller input for equip to specify tap so that hold to drop functions properly.