Projectile Weapon Functional
Projectiles spawned from a ranged weapon are successfully obtaining information about the weapon and character who shot them.

They are also successfully calling the GetHit function on the enemy on contact to cause damage.

Next need to test and adjust the bullet size and velocity to make it play well.