Adjustments to Speed/Size of Projectile
I made the actual projectile invisible and larger. And I added a smaller visible mesh. This makes hit detection more forgiving. Will adjust.

Spawn position of the projectile, and impulse direction adjusted to account for the direction player is facing.

Impulse increased to speed up the projectile.

TODO: Speed up the animation. Try to adjust he animation so the weapon is more centered. Use a different running animation so the weapon is more centered.