Enemy Weapon Attached to Socket
Set up enemy weapon attach to socket based on enum WeaponType.
Uses combination of blueprints BP_Enemy and Weapon.cpp + Enemy.cpp to get WeaponType, get the socket name, attach the correct weapon to the socket.
The BP for a weapon can be attached to an individual enemy to override default, or a default BP_weapon will be picked based on WeaponType if it is null.

TODO: Next going through videos to set up weapon attacks