Character Get Hit Functionality in Progress
Working on implementing AI for the enemy to attack the character and the character to react by calling their get hit method.
Enemy attacks are happening but need to set the timer between attacks to zero when enemy first reaches character and need enemy to rotate to face character.
Character get hit is getting called but it is unreliable. Needs to be fixed.
Weapon socketing in enemy hand is bugged. It is not quite attached and as it hits character it is being pushed further out of socket. Need to fix.
TODO: need to set this aside and go through guides to save time on fixing these issues. come back to stuff specific to my project after all videos done in section 18.