18.185
Added a follow distance property to weapons, this is checked by the enemy to set the distance to follow dynamically.

Integrated CombatMode Enum with enemy attacks and death condition for both enemy and player. For enemy this is set to chasing after attack anim ends.

Added an abort attack method which is called on GetHit. This fixes a bug when Characters could end up stuck in Attacking ActionState if hit mid animation (because the animation never ended and called animnotify attackend).

Refactored the logic regarding when an enemy should attack. Added the condition that the enemy should be chasing to attack (thus not mid attack animation).