Working on RootMotion Attacks
Echo is warping to the target but after the attacks she is rotating upward. This bug is only happening if rootmotion is turned on, but without root motion we can't warp to target with motion warping.

TODO: turn off motionwarping for Echo for now. Come back to it later maybe. Move on to section 20 next.