Potion Functionality Progress
Logic for Recovery potions (which give X amount of health) is in place.

Buff Logic is in progress. Have a switch system which turns on the bools and multipliers. And a timer which then starts and turns then off after buff duration passes.

TODO: use the bools and multipliers in the appropriate places in the code so that the buffs will take effect. And then create blueprints for each buff type with different item meshes or niagara effects like the soul effect. Then add them as possible drops from destructibles.