Potion and Buff UE5 Assets and Popup Text
Created blueprints for each potion and buff type with enums and values set and an in game visual.

Added a text box to the overlay and setup functionality for it to display the type of buff or potion briefly.

TODO: 1. use the bools and multipliers in the appropriate places in the code so that the buffs will take effect. 2. Add the potions and buffs as possible drops from destructibles.