ABP Enemy Template
ABP Enemy Template created and applied to Alien Solider.
Refactored CombatMode setting to go through a setter function to help debug issues.
Disabled Soul Value setting in Enemy.cpp because it is causing a UE5 crash.
Enemies are looping death animation. Find the cause and fix this. (likely it is the asset set in ABP_Enemy to the template.)

//TODO
1. Stop enemy death anim from looping.
2. Continue through section 22.
3.  debug soul value setting issue
4. fix root motion animations. still causing echo to rotated off ground.